Classes re-introduced successfully
Classes are now their own proper file; shield system now in place.
